Extract buffered event XDC test helpers
