+++ tests/xdc/buffered_nexus_events_replication_test.go
@@ -10,26 +10,13 @@ import (
 	commandpb "go.temporal.io/api/command/v1"
 	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
-	historypb "go.temporal.io/api/history/v1"
-	nexuspb "go.temporal.io/api/nexus/v1"
-	"go.temporal.io/api/operatorservice/v1"
 	taskqueuepb "go.temporal.io/api/taskqueue/v1"
 	"go.temporal.io/api/workflowservice/v1"
-	"go.temporal.io/server/common"
 	commonnexus "go.temporal.io/server/common/nexus"
-	"go.temporal.io/server/common/nexus/nexusrpc"
 	"go.temporal.io/server/common/nexus/nexustest"
 	"go.temporal.io/server/common/testing/await"
-	"go.temporal.io/server/components/nexusoperations"
-	"go.temporal.io/server/tests/testcore"
 	"google.golang.org/protobuf/types/known/durationpb"
 )
-
-type bufferedNexusCallback struct {
-	operation string
-	url       string
-	token     string
-}
 
 // TestBufferedNexusEventsReapplySharedOperationAndSkipLosingOnlyOperation buffers completion of an
 // operation shared by both branches plus start and completion of an operation created only on the
@@ -402,218 +389,4 @@ func (s *NexusStateReplicationSuite) TestNaturallyBufferedNexusCancelRequestComp
 		hasNexusEventForScheduledID(winningHistory, enumspb.EVENT_TYPE_NEXUS_OPERATION_CANCEL_REQUEST_COMPLETED, scheduledID),
 		"Nexus cancellation results are not cherry-pickable and must remain only on the losing branch",
 	)
-}
-
-func (s *NexusStateReplicationSuite) createBufferedNexusEndpoint(ctx context.Context, handler nexustest.Handler) string {
-	s.T().Helper()
-	listenAddress := nexustest.AllocListenAddress()
-	nexustest.NewNexusServer(s.T(), listenAddress, handler)
-	for _, cluster := range s.clusters {
-		cluster.OverrideDynamicConfig(
-			s.T(),
-			nexusoperations.CallbackURLTemplate,
-			"http://"+s.clusters[0].Host().FrontendHTTPAddress()+"/namespaces/{{.NamespaceName}}/nexus/callback",
-		)
-	}
-	endpointName := testcore.RandomizedNexusEndpoint(s.T().Name())
-	for _, cluster := range s.clusters {
-		_, err := cluster.OperatorClient().CreateNexusEndpoint(ctx, &operatorservice.CreateNexusEndpointRequest{
-			Spec: &nexuspb.EndpointSpec{
-				Name: endpointName,
-				Target: &nexuspb.EndpointTarget{Variant: &nexuspb.EndpointTarget_External_{
-					External: &nexuspb.EndpointTarget_External{Url: "http://" + listenAddress},
-				}},
-			},
-		})
-		s.Require().NoError(err)
-	}
-	return endpointName
-}
-
-func (s *NexusStateReplicationSuite) failNexusOperation(ctx context.Context, callback bufferedNexusCallback) {
-	client := nexusrpc.NewCompletionHTTPClient(nexusrpc.CompletionHTTPClientOptions{Serializer: commonnexus.PayloadSerializer})
-	err := client.CompleteOperation(ctx, callback.url, nexusrpc.CompleteOperationOptions{
-		Error: &nexus.OperationError{
-			State: nexus.OperationStateFailed,
-			Cause: &nexus.FailureError{Failure: nexus.Failure{Message: "expected operation failure"}},
-		},
-		Header: nexus.Header{commonnexus.CallbackTokenHeader: callback.token},
-	})
-	s.Require().NoError(err)
-}
-
-func (s *NexusStateReplicationSuite) setupBufferedNexusEndpoint(
-	ctx context.Context,
-) (string, <-chan bufferedNexusCallback, chan struct{}) {
-	s.T().Helper()
-	operationCallbacks := make(chan bufferedNexusCallback, 2)
-	allowLosingOnlyOperationStart := make(chan struct{})
-	handler := nexustest.Handler{
-		OnStartOperation: func(
-			_ context.Context,
-			_, operation string,
-			_ *nexus.LazyValue,
-			options nexus.StartOperationOptions,
-		) (nexus.HandlerStartOperationResult[any], error) {
-			operationCallbacks <- bufferedNexusCallback{
-				operation: operation,
-				url:       options.CallbackURL,
-				token:     options.CallbackHeader.Get(commonnexus.CallbackTokenHeader),
-			}
-			if operation == "losing-only-operation" {
-				select {
-				case <-allowLosingOnlyOperationStart:
-				case <-ctx.Done():
-					return nil, ctx.Err()
-				}
-			}
-			return &nexus.HandlerStartOperationResultAsync{OperationToken: operation}, nil
-		},
-	}
-	listenAddress := nexustest.AllocListenAddress()
-	nexustest.NewNexusServer(s.T(), listenAddress, handler)
-	for _, cluster := range s.clusters {
-		cluster.OverrideDynamicConfig(
-			s.T(),
-			nexusoperations.CallbackURLTemplate,
-			"http://"+s.clusters[0].Host().FrontendHTTPAddress()+"/namespaces/{{.NamespaceName}}/nexus/callback",
-		)
-	}
-
-	endpointName := testcore.RandomizedNexusEndpoint(s.T().Name())
-	for _, cluster := range s.clusters {
-		_, err := cluster.OperatorClient().CreateNexusEndpoint(ctx, &operatorservice.CreateNexusEndpointRequest{
-			Spec: &nexuspb.EndpointSpec{
-				Name: endpointName,
-				Target: &nexuspb.EndpointTarget{Variant: &nexuspb.EndpointTarget_External_{
-					External: &nexuspb.EndpointTarget_External{Url: "http://" + listenAddress},
-				}},
-			},
-		})
-		s.Require().NoError(err)
-	}
-	return endpointName, operationCallbacks, allowLosingOnlyOperationStart
-}
-
-func scheduleBufferedNexusOperationCommand(endpoint, operation string) *commandpb.Command {
-	return scheduleBufferedNexusOperationCommandWithTimeout(endpoint, operation, time.Minute)
-}
-
-func scheduleBufferedNexusOperationCommandWithTimeout(endpoint, operation string, timeout time.Duration) *commandpb.Command {
-	return &commandpb.Command{
-		CommandType: enumspb.COMMAND_TYPE_SCHEDULE_NEXUS_OPERATION,
-		Attributes: &commandpb.Command_ScheduleNexusOperationCommandAttributes{
-			ScheduleNexusOperationCommandAttributes: &commandpb.ScheduleNexusOperationCommandAttributes{
-				Endpoint:               endpoint,
-				Service:                "service",
-				Operation:              operation,
-				ScheduleToCloseTimeout: durationpb.New(timeout),
-			},
-		},
-	}
-}
-
-func requestCancelNexusOperationCommand(scheduledEventID int64) *commandpb.Command {
-	return &commandpb.Command{
-		CommandType: enumspb.COMMAND_TYPE_REQUEST_CANCEL_NEXUS_OPERATION,
-		Attributes: &commandpb.Command_RequestCancelNexusOperationCommandAttributes{
-			RequestCancelNexusOperationCommandAttributes: &commandpb.RequestCancelNexusOperationCommandAttributes{
-				ScheduledEventId: scheduledEventID,
-			},
-		},
-	}
-}
-
-func receiveBufferedNexusCallback(
-	ctx context.Context,
-	t require.TestingT,
-	operationCallbacks <-chan bufferedNexusCallback,
-	expectedOperation string,
-) bufferedNexusCallback {
-	select {
-	case operationCallback := <-operationCallbacks:
-		require.Equal(t, expectedOperation, operationCallback.operation)
-		return operationCallback
-	case <-ctx.Done():
-		require.FailNow(t, "timed out waiting for Nexus operation callback", expectedOperation)
-		return bufferedNexusCallback{}
-	}
-}
-
-func receiveAnyBufferedNexusCallback(
-	ctx context.Context,
-	t require.TestingT,
-	operationCallbacks <-chan bufferedNexusCallback,
-) bufferedNexusCallback {
-	select {
-	case operationCallback := <-operationCallbacks:
-		return operationCallback
-	case <-ctx.Done():
-		require.FailNow(t, "timed out waiting for Nexus operation callback")
-		return bufferedNexusCallback{}
-	}
-}
-
-func (s *NexusStateReplicationSuite) findBufferedNexusLosingBranch(
-	ctx context.Context,
-	ns string,
-	namespaceID string,
-	execution *commonpb.WorkflowExecution,
-	losingScheduledID int64,
-) []*historypb.HistoryEvent {
-	s.T().Helper()
-	history := s.findNonCurrentHistoryBranch(ctx, ns, namespaceID, execution, func(history []*historypb.HistoryEvent) bool {
-		return hasNexusEventForScheduledID(history, enumspb.EVENT_TYPE_NEXUS_OPERATION_COMPLETED, losingScheduledID)
-	})
-	for _, event := range history {
-		if event.EventType == enumspb.EVENT_TYPE_NEXUS_OPERATION_STARTED ||
-			event.EventType == enumspb.EVENT_TYPE_NEXUS_OPERATION_COMPLETED {
-			s.Require().NotEqual(common.BufferedEventID, event.EventId)
-			s.Require().Positive(event.EventId)
-		}
-	}
-	return history
-}
-
-func findNexusScheduledEventID(history []*historypb.HistoryEvent, operation string) int64 {
-	for _, event := range history {
-		if event.EventType == enumspb.EVENT_TYPE_NEXUS_OPERATION_SCHEDULED &&
-			event.GetNexusOperationScheduledEventAttributes().Operation == operation {
-			return event.EventId
-		}
-	}
-	return common.EmptyEventID
-}
-
-func hasNexusEventForScheduledID(history []*historypb.HistoryEvent, eventType enumspb.EventType, scheduledEventID int64) bool {
-	for _, event := range history {
-		if event.EventType != eventType {
-			continue
-		}
-		var eventScheduledID int64
-		switch eventType {
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_SCHEDULED:
-			eventScheduledID = event.EventId
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_STARTED:
-			eventScheduledID = event.GetNexusOperationStartedEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_COMPLETED:
-			eventScheduledID = event.GetNexusOperationCompletedEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_FAILED:
-			eventScheduledID = event.GetNexusOperationFailedEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_CANCELED:
-			eventScheduledID = event.GetNexusOperationCanceledEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_TIMED_OUT:
-			eventScheduledID = event.GetNexusOperationTimedOutEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_CANCEL_REQUEST_COMPLETED:
-			eventScheduledID = event.GetNexusOperationCancelRequestCompletedEventAttributes().GetScheduledEventId()
-		case enumspb.EVENT_TYPE_NEXUS_OPERATION_CANCEL_REQUEST_FAILED:
-			eventScheduledID = event.GetNexusOperationCancelRequestFailedEventAttributes().GetScheduledEventId()
-		default:
-			return false
-		}
-		if eventScheduledID == scheduledEventID {
-			return true
-		}
-	}
-	return false
 }
